Refactor handling of Java types.
Replaces CompilerState.getClass by several methods, which are more
specific about what Java type they return (the unparameterized class,
the parameterized class, the primitive type, etc.).

BUG=
R=springerm@google.com

Review URL: https://chromereviews.googleplex.com//499857013 .

+++ test/unit/dynamic/scenario.dart
@@ -27,7 +27,7 @@ int testDynamicDispatchPrimitive() {
 
   dynamic v3 = 10.5;
   dynamic v4 = 20.5;
-  result = result + (v3 + v4) as int;
+  result = (result + (v3 + v4) as double).toInt();
 
   return result;
 }
